refactor: log module
